ml_core/requirements.txt
sentence-transformers>=2.7.0
faiss-cpu>=1.10.0
pandas>=2.0.0
numpy>=1.24.0